Bug fixes.
Currently it trips over Define Sprite tags, so implementing those is the next step.
The scope_stack is not enough because the length is determined by the size of the tag and the current implementation would skip over that.
And although spec mandated, many implementations of DefineSprite do no conclude it with an End Tag, so that has to be it's own peculiarity.