Tag counting and id interface added
Conflicting ids is a serious issue without a definite or sensible resolution, thus it's an error and not a peculiarity. It's very easily recoverable from however, just make a choice whether the existing id should remain or if the tag at the end of the tag_stream should be chosen in it's stead.